requirements.txt
# Forecasting Project Dependencies
# Python 3.13.5 compatible

# Data manipulation and analysis
pandas>=2.1.0
numpy>=1.24.0

# ===== NOUVEAU : SQL Server Support =====
# Driver ODBC pour SQL Server
pyodbc>=5.0.0

# Gestion des variables d'environnement (.env)
python-dotenv>=1.0.0

# Database abstraction (optionnel, pour futures extensions)
sqlalchemy>=2.0.0

# ===== Visualization =====
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ===== Machine Learning =====
scikit-learn>=1.3.0

# ===== Statistical analysis =====
scipy>=1.11.0
statsmodels>=0.14.0

# ===== Testing =====
pytest>=7.4.0
pytest-cov>=4.1.0

# ===== Data validation =====
pydantic>=2.0.0

# ===== Utilities =====
python-dateutil>=2.8.0
tqdm>=4.65.0  # Progress bars

# ===== Development tools =====
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting

# ===== Documentation =====
sphinx>=7.0.0  # Génération de la doc automatique

# Pour SQL Server :
# Windows :
#   1. Installer le driver ODBC : https://learn.microsoft.com/en-us/sql/connect/odbc/download-odbc-driver-for-sql-server
#   2. pip install pyodbc

# Linux :
#   1. sudo apt-get install unixodbc-dev
#   2. pip install pyodbc

# macOS :
#   1. brew install unixodbc
#   2. pip install pyodbc

# Vérifier les drivers disponibles :
#   python -c "import pyodbc; print(pyodbc.drivers())"